# src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Learning Management System Course APIs",
		"description": "This is a Learning Management System API documentation which is only exposed on development environment.",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "https://api.lms-staging.com/api/main",
			"description": "LMS API's related to courses"
		},
		{
			"url": "http://localhost/api/main",
			"description": "Local host server, exclusively for back-end developers."
		}
	],
	"paths": {
		"/course/upload-image": {
			"post": {
				"summary": "Upload a course image",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"mimeType": {
										"type": "string",
										"enum": [
											"image/jpg",
											"image/jpeg",
											"image/png"
										]
									}
								},
								"example": {
									"mimeType": "image/jpg"
								},
								"required": [
									"mimeType"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"url": {
													"type": "string"
												},
												"fields": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"key": {
																"type": "string"
															},
															"bucket": {
																"type": "string"
															},
															"X-Amz-Algorithm": {
																"type": "string"
															},
															"X-Amz-Credential": {
																"type": "string"
															},
															"X-Amz-Date": {
																"type": "string"
															},
															"Policy": {
																"type": "string"
															},
															"X-Amz-Signature": {
																"type": "string"
															}
														},
														"required": [
															"key",
															"bucket",
															"X-Amz-Algorithm",
															"X-Amz-Credential",
															"X-Amz-Date",
															"Policy",
															"X-Amz-Signature"
														]
													}
												}
											},
											"required": [
												"url",
												"fields"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/course/upload-lecture-video": {
			"post": {
				"summary": "Upload lecture video",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"mimeType": {
										"type": "string",
										"enum": [
											"video/mp4"
										]
									}
								},
								"example": {
									"mimeType": "video/mp4"
								},
								"required": [
									"mimeType"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"url": {
													"type": "string"
												},
												"fields": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"key": {
																"type": "string"
															},
															"bucket": {
																"type": "string"
															},
															"X-Amz-Algorithm": {
																"type": "string"
															},
															"X-Amz-Credential": {
																"type": "string"
															},
															"X-Amz-Date": {
																"type": "string"
															},
															"Policy": {
																"type": "string"
															},
															"X-Amz-Signature": {
																"type": "string"
															}
														},
														"required": [
															"key",
															"bucket",
															"X-Amz-Algorithm",
															"X-Amz-Credential",
															"X-Amz-Date",
															"Policy",
															"X-Amz-Signature"
														]
													}
												}
											},
											"required": [
												"url",
												"fields"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/course/upload-lecture-subtitle": {
			"post": {
				"summary": "Upload lecture subtitle",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"mimeType": {
										"type": "string",
										"enum": [
											"text/vtt"
										]
									}
								},
								"example": {
									"mimeType": "text/vtt"
								},
								"required": [
									"mimeType"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"url": {
													"type": "string"
												},
												"fields": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"key": {
																"type": "string"
															},
															"bucket": {
																"type": "string"
															},
															"X-Amz-Algorithm": {
																"type": "string"
															},
															"X-Amz-Credential": {
																"type": "string"
															},
															"X-Amz-Date": {
																"type": "string"
															},
															"Policy": {
																"type": "string"
															},
															"X-Amz-Signature": {
																"type": "string"
															}
														},
														"required": [
															"key",
															"bucket",
															"X-Amz-Algorithm",
															"X-Amz-Credential",
															"X-Amz-Date",
															"Policy",
															"X-Amz-Signature"
														]
													}
												}
											},
											"required": [
												"url",
												"fields"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/course/create-by-instructor": {
			"post": {
				"summary": "Create course by instructor",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string",
										"example": "The Complete Python Bootcamp From Zero to Hero in Python"
									},
									"description": {
										"type": "string",
										"example": "Learn Python like a Professional Start from the basics and go all the way to creating your own applications and games"
									},
									"category": {
										"type": "string",
										"example": "PROGRAMMING"
									},
									"languages": {
										"type": "array",
										"items": {
											"type": "string",
											"enum": [
												"English"
											],
											"example": "English"
										}
									},
									"subtitles": {
										"type": "array",
										"items": {
											"type": "string",
											"enum": [
												"English",
												"French",
												"Germany"
											],
											"example": "English"
										}
									},
									"learnings": {
										"type": "array",
										"items": {
											"type": "string",
											"example": "You will learn how to leverage the power of Python to solve tasks."
										}
									},
									"materialsAndOffers": {
										"type": "array",
										"items": {
											"type": "string",
											"example": "22 hours on-demand video"
										}
									},
									"price": {
										"type": "object",
										"properties": {
											"currency": {
												"type": "string",
												"enum": [
													"CAD"
												],
												"example": "CAD"
											},
											"value": {
												"type": "number",
												"example": 100
											}
										},
										"required": [
											"currency",
											"value"
										]
									},
									"image": {
										"type": "string",
										"example": "https://s3.us-east-1.amazonaws.com/learning-management-system-project/public/courses/images/37f3aa96-7968-428f-9203-67a0a4c9d7e0.jpg"
									},
									"sections": {
										"type": "array",
										"items": {
											"type": "object",
											"properties": {
												"title": {
													"type": "string",
													"example": "Course Overview"
												},
												"lectures": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"title": {
																"type": "string",
																"example": "Auto-Welcome Message"
															},
															"description": {
																"type": "string",
																"example": "Welcome to the Complete Python Bootcamp"
															},
															"link": {
																"type": "string",
																"example": "https://s3.us-east-1.amazonaws.com/learning-management-system-project/public/courses/raw-lectures/a4f8287d-2fa5-44c9-9223-4a2b292b98ad.mp4"
															},
															"subtitles": {
																"type": "array",
																"items": {
																	"type": "object",
																	"properties": {
																		"language": {
																			"type": "string",
																			"enum": [
																				"English",
																				"French",
																				"Germany"
																			],
																			"example": "English"
																		},
																		"url": {
																			"type": "string",
																			"example": "https://s3.us-east-1.amazonaws.com/learning-management-system-project/public/courses/raw-lectures/84036c24-17ca-4405-bbfd-ffb679ceb9d0.srt"
																		}
																	},
																	"required": [
																		"language",
																		"url"
																	]
																}
															},
															"order": {
																"type": "number",
																"example": 1
															}
														},
														"required": [
															"title",
															"description",
															"link",
															"subtitles",
															"order"
														]
													}
												},
												"order": {
													"type": "number",
													"example": 1
												}
											},
											"required": [
												"title",
												"lectures",
												"order"
											]
										}
									}
								},
								"required": [
									"title",
									"description",
									"category",
									"languages",
									"subtitles",
									"learnings",
									"materialsAndOffers",
									"price",
									"image",
									"sections"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"$ref": "#/components/schemas/Course"
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
...
		"/course/my-learnings": {
			"get": {
				"summary": "Get the list of enrolled courses of a student",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Course"
											}
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/course/my-course": {
			"get": {
				"summary": "Get the enrolled course of a student",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "courseId",
						"in": "query",
						"description": "Course id",
						"required": true,
						"schema": {
							"type": "string",
							"example": "667f14303cebd3afba47d623"
						}
					}
				],
				"requestBody": {},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"$ref": "#/components/schemas/Course"
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
...
		"/course/all-by-instructor": {
			"get": {
				"summary": "Get the list of courses created by instructor",
				"tags": [
					"Course"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Course"
											}
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
...
	}
}